requirements.txt
openai
oss2
python-dotenv
gradio
tenacity
elevenlabs